pyproject.toml
[build-system]
requires = [
    "setuptools>=80",
    "wheel>=0.45.1",
    "setuptools_scm[toml]>=8.3.0",
]
build-backend = "setuptools.build_meta"

[project]
name = "c3a3-config_injector"
description = "Configuration Wrapping Framework - Declarative YAML specs for wrapping executables with configuration injection"
readme = { file = "README.md", content-type = "text/markdown" }
requires-python = ">=3.11"
authors = [{ name = "ã", email = "me@c3a3.me" }]
license = { file = "LICENSE" }
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: Other/Proprietary License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: System :: Systems Administration",
]
dependencies = [
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "types-PyYAML",
    "python-dotenv>=1.0.0",
    "typer>=0.9.0",
    "rich>=13.0.0",
]
dynamic = ["version"]

[project.optional-dependencies]
dev = [
    "pytest>=8",
    "pytest-cov",
    "pytest-approvaltests>=0.2.4",
    "mypy",
    "ruff",
    "black>=23.0.0",
    "pre-commit>=3.0.0",
]
bws = [
    "bitwarden-sdk>=1.0.0",
]

[project.scripts]
wrapper = "config_injector.cli:app"

[project.urls]
Homepage = "https://github.com/mrtu/c3a3-config_injector"
Documentation = "https://github.com/mrtu/c3a3-config_injector#readme"
Issues = "https://github.com/mrtu/c3a3-config_injector/issues"
"Bug Tracker" = "https://github.com/mrtu/c3a3-config_injector/issues"
"License" = "https://github.com/mrtu/c3a3-config_injector/blob/main/LICENSE"

[tool.setuptools]
package-dir = { "" = "src" }
include-package-data = true

[tool.setuptools.packages.find]
where = ["src"]
exclude = ["tests*"]

[tool.setuptools_scm]
fallback_version = "1.0.0"
tag_regex = "^(?P<version>\\d+\\.\\d+\\.\\d+)$"

[tool.black]
line-length = 88
target-version = ['py311']

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "B", "C4", "UP", "ARG", "SIM", "TCH", "TID", "Q"]
ignore = ["E501", "B008"]

[tool.mypy]
strict = true
files = ["src/config_injector"]
exclude = ["tests/.*"]

[tool.pytest.ini_options]
addopts = "-ra -q --cov=src/config_injector --cov-report=term-missing --cov-report=html"
testpaths = ["tests"] 
